move git functions to git package
