+++ git/git.go
@@ -1,11 +1,14 @@
 package git
 
 import (
+	"bytes"
 	"context"
+	"os/exec"
+	"path"
+	"path/filepath"
 	"strings"
 	"time"
 
-	"github.com/cdnjs/tools/packages"
 	"github.com/cdnjs/tools/util"
 )
 
@@ -24,7 +27,7 @@ func (g Version) Get() string {
 // Download will git check out a particular version.
 func (g Version) Download(args ...interface{}) string {
 	ctx, dir := args[0].(context.Context), args[1].(string)
-	packages.GitForceCheckout(ctx, dir, g.Tag)
+	ForceCheckout(ctx, dir, g.Tag)
 	return dir // download dir is the same as original dir
 }
 
@@ -39,8 +42,8 @@ func (g Version) GetTimeStamp() time.Time {
 
 // GetVersions gets all of the versions associated with a git repo,
 // as well as the latest version.
-func GetVersions(ctx context.Context, pckg *packages.Package, packageGitcache string) ([]Version, *string) {
-	gitTags := packages.GitTags(ctx, packageGitcache)
+func GetVersions(ctx context.Context, packageGitcache string) ([]Version, *string) {
+	gitTags := Tags(ctx, packageGitcache)
 	util.Debugf(ctx, "found tags in git: %s\n", gitTags)
 
 	gitVersions := make([]Version, 0)
@@ -49,7 +52,7 @@ func GetVersions(ctx context.Context, pckg *packages.Package, packageGitcache st
 		gitVersions = append(gitVersions, Version{
 			Tag:       tag,
 			Version:   version,
-			TimeStamp: packages.GitTimeStamp(ctx, packageGitcache, tag),
+			TimeStamp: TimeStamp(ctx, packageGitcache, tag),
 		})
 	}
 
@@ -57,4 +60,176 @@ func GetVersions(ctx context.Context, pckg *packages.Package, packageGitcache st
 		return gitVersions, &latest.Version
 	}
 	return gitVersions, nil
+}
+
+// ListPackageVersions first lists all the versions (and top-level package.json)
+// in the package and passes the list to git ls-tree which filters out
+// those not in the tree.
+func ListPackageVersions(ctx context.Context, basePath string) []string {
+	filesOnFs, err := filepath.Glob(path.Join(basePath, "*"))
+	util.Check(err)
+
+	filteredFilesOnFs := make([]string, 0)
+
+	// filter out package.json
+	for _, file := range filesOnFs {
+		if !strings.HasSuffix(file, ".donotoptimizepng") && !strings.HasSuffix(file, "package.json") && strings.Trim(file, " ") != "" {
+			filteredFilesOnFs = append(filteredFilesOnFs, file)
+		}
+	}
+
+	// no local version, no need to check what's in git
+	if len(filteredFilesOnFs) == 0 {
+		return make([]string, 0)
+	}
+
+	args := []string{
+		"ls-tree", "--name-only", "origin/master",
+	}
+	args = append(args, filteredFilesOnFs...)
+
+	cmd := exec.Command("git", args...)
+	cmd.Dir = basePath
+	util.Debugf(ctx, "run %s from %s\n", cmd, basePath)
+	out := util.CheckCmd(cmd.CombinedOutput())
+
+	outFiles := strings.Split(out, "\n")
+
+	filteredOutFiles := make([]string, 0)
+	// remove basePath from the output
+	for _, v := range outFiles {
+		if strings.Trim(v, " ") != "" {
+			filteredOutFiles = append(
+				filteredOutFiles, strings.ReplaceAll(v, basePath+"/", ""))
+		}
+	}
+
+	if util.IsDebug() {
+		diff := arrDiff(filteredFilesOnFs, outFiles)
+		if len(diff) > 0 {
+			util.Printf(ctx, "found %d staged versions\n", len(diff))
+		}
+	}
+
+	return filteredOutFiles
+}
+
+// Add adds to the next commit.
+func Add(ctx context.Context, gitpath, relpath string) {
+	args := []string{"add", relpath}
+
+	cmd := exec.Command("git", args...)
+	cmd.Dir = gitpath
+	util.Debugf(ctx, "run %s\n", cmd)
+	util.CheckCmd(cmd.CombinedOutput())
+}
+
+// Commit makes a new commit.
+func Commit(ctx context.Context, gitpath, msg string) {
+	args := []string{"commit", "-m", msg}
+
+	cmd := exec.Command("git", args...)
+	cmd.Dir = gitpath
+	util.Debugf(ctx, "run %s\n", cmd)
+	util.CheckCmd(cmd.CombinedOutput())
+}
+
+// Fetch fetches objs/refs to the repository.
+func Fetch(ctx context.Context, gitpath string) ([]byte, error) {
+	args := []string{"fetch"}
+
+	cmd := exec.Command("git", args...)
+	cmd.Dir = gitpath
+	util.Debugf(ctx, "%s: run %s\n", gitpath, cmd)
+	return cmd.CombinedOutput()
+}
+
+// Push pushes to a git repository.
+func Push(ctx context.Context, gitpath string) {
+	args := []string{"push"}
+
+	cmd := exec.Command("git", args...)
+	cmd.Dir = gitpath
+	util.Debugf(ctx, "run %s\n", cmd)
+	util.CheckCmd(cmd.CombinedOutput())
+}
+
+// Clone clones a git repository.
+func Clone(ctx context.Context, target string, gitpath string) ([]byte, error) {
+	args := []string{"clone", target, "."}
+
+	cmd := exec.Command("git", args...)
+	cmd.Dir = gitpath
+	util.Debugf(ctx, "%s: run %s\n", gitpath, cmd)
+	out, err := cmd.CombinedOutput()
+	return out, err
+}
+
+// Tags returns the []string of git tags for a package.
+func Tags(ctx context.Context, gitpath string) []string {
+	args := []string{"tag"}
+
+	cmd := exec.Command("git", args...)
+	cmd.Dir = gitpath
+	util.Debugf(ctx, "run %s\n", cmd)
+	out := util.CheckCmd(cmd.CombinedOutput())
+
+	tags := make([]string, 0)
+	for _, line := range strings.Split(out, "\n") {
+		if strings.Trim(line, " ") != "" {
+			tags = append(tags, line)
+		}
+	}
+
+	return tags
+}
+
+// TimeStamp gets the time stamp for a particular tag (ex. v1.0).
+func TimeStamp(ctx context.Context, gitpath, tag string) time.Time {
+	args := []string{"log", "-1", "--format=%aI", tag}
+
+	var out bytes.Buffer
+	var stderr bytes.Buffer
+
+	cmd := exec.Command("git", args...)
+	cmd.Dir = gitpath
+	cmd.Stdout = &out
+	cmd.Stderr = &stderr
+	util.Debugf(ctx, "run %s\n", cmd)
+	err := cmd.Run()
+
+	if err != nil {
+		util.Errf(ctx, "%s: %s\n", err, stderr.String())
+		return time.Unix(0, 0)
+	}
+
+	t, err := time.Parse(time.RFC3339, strings.TrimSpace(out.String()))
+	util.Check(err)
+
+	return t
+}
+
+// ForceCheckout force checkouts a particular tag.
+func ForceCheckout(ctx context.Context, gitpath, tag string) {
+	args := []string{"checkout", tag, "-f"}
+
+	cmd := exec.Command("git", args...)
+	cmd.Dir = gitpath
+	util.Debugf(ctx, "run %s\n", cmd)
+	util.CheckCmd(cmd.CombinedOutput())
+}
+
+func arrDiff(a, b []string) (diff []string) {
+	m := make(map[string]bool)
+
+	for _, item := range b {
+		m[item] = true
+	}
+
+	for _, item := range a {
+		if _, ok := m[item]; !ok {
+			diff = append(diff, item)
+		}
+	}
+	return
 }
