+++ git/github.go
@@ -1,12 +1,17 @@
 package git
 
 import (
+	"context"
 	"encoding/json"
 	"io/ioutil"
 	"net/http"
 	"regexp"
 
 	"github.com/cdnjs/tools/util"
+
+	githubapi "github.com/google/go-github/github"
+	"golang.org/x/oauth2"
+	githuboauth2 "golang.org/x/oauth2/github"
 )
 
 // Stars holds the number of stars for a GitHub repository.
@@ -34,5 +39,17 @@ func GetGitHubStars(gitURL string) Stars {
 	var stars Stars
 	util.Check(json.Unmarshal(body, &stars))
 	return stars
+}
 
+// GetClient gets a GitHub client to interact with its API.
+func GetClient() *githubapi.Client {
+	token := util.GetEnv("GITHUB_REPO_API_KEY")
+	ctx := context.Background()
+	conf := &oauth2.Config{
+		Endpoint: githuboauth2.Endpoint,
+	}
+
+	tc := conf.Client(ctx, &oauth2.Token{AccessToken: token})
+
+	return githubapi.NewClient(tc)
 }
